fix: resolve 17 Klocwork violations across all device families
Addresses 17 active Klocwork violations identified in the
PMIC LLD codebase:

MISRA.DEFINE.NOT_DISTINCT.C90.2012 - MISRA Required
   - TPS65036x-Q1: Shortened NINT_GPI identifier names (LPM_CTRL_MODE_INPUT → LPM_INPUT)
   - TPS6522x-Q1: Shortened EN_PB_VSENSE identifier names (FXN_SEL → FN)
   - TPS65386x-Q1: Shortened PLDO mode names (EN_AS_LDO_IN_OPER → EN_LDO_OPER)

INVARIANT_CONDITION.GEN Pattern A - MISRA Required
   - TPS6522x-Q1/pmic_fsm.c: Removed redundant status checks after early returns
     in Pmic_fsmSetTriggerCfg() and Pmic_fsmGetTriggerCfg()
   - TPS6522x-Q1/pmic_gpio.c: Removed redundant check in Pmic_gpioGetEnPbVSenseStatus()
   - TPS65386x-Q1/pmic_irq.c: Removed redundant checks in Pmic_irqSetCfgs() loops

MISRA.ETYPE.INAPPR.CAST.2012 - MISRA Advisory
   - TPS6522x-Q1/pmic_wdg.c: Fixed inappropriate bool casts (uint8_t → bool)
     using explicit comparison (mode != 0U) instead of (bool)mode
   - TPS65386x-Q1/pmic_core.c: Fixed bool-to-uint8_t cast using ternary operator
     following established pattern in same file

CWARN.BITOP.SIZE - Review
   - TPS6522x-Q1/pmic_irq.c: Fixed bitwise operation type mismatch
     Changed (uint8_t)(1U << bitIndex) to (uint32_t)(1U << bitIndex)
     to match intrStat[] array element type

MISRA.PPARAM.NEEDS.CONST - MISRA Advisory
   - TPS65386x-Q1/pmic.c: Added const qualifier to validateComms() parameter
     Function only reads handle, doesn't modify it

+++ devices/TPS6522x-Q1/src/pmic_irq.c
@@ -328,7 +328,7 @@ int32_t Pmic_irqGetStatus(const Pmic_Handle_t *handle, Pmic_IrqStatus_t *irqStat
                 // Set the corresponding bit in the status array
                 uint8_t arrayIndex = irqNum / 32U;
                 uint8_t bitIndex = irqNum % 32U;
-                irqStatLocal.intrStat[arrayIndex] |= (uint8_t)(1U << bitIndex);
+                irqStatLocal.intrStat[arrayIndex] |= ((uint32_t)1U << bitIndex);
             }
         }
     }
